Transfer-api - Adiciona aorquestração para a realização de transferências entre contas

+++ src/Transfer-Api/Transfer.Api/Controllers/Requests/CreateTransferRequest.cs
@@ -1,3 +1,3 @@
 ﻿namespace Transfer.Api.Controllers.Requests;
 
-public record CreateTransferRequest(decimal DestinationAccountNumber, decimal Amount);
+public record CreateTransferRequest(int DestinationAccountNumber, decimal Amount);
